refactor: Add cdc/core entry point and update configs.
