+++ integration-libs/cdc/core/public_api.ts
@@ -1,0 +1,4 @@
+/*
+ * Public API Surface for CDC
+ */
+export * from './index';
