+++ projects/storefrontapp/tsconfig.app.prod.json
@@ -104,6 +104,7 @@
       "@spartacus/tracking/tms/gtm": ["dist/tracking/tms/gtm"],
       "@spartacus/tracking/tms": ["dist/tracking/tms"],
       "@spartacus/cdc/components": ["dist/cdc/components"],
+      "@spartacus/cdc/core": ["dist/cdc/core"],
       "@spartacus/cdc": ["dist/cdc"],
       "@spartacus/cds": ["dist/cds"],
       "@spartacus/assets": ["dist/assets"],
