+++ tsconfig.compodoc.json
@@ -194,6 +194,9 @@
       "@spartacus/cdc/components": [
         "integration-libs/cdc/components/public_api"
       ],
+      "@spartacus/cdc/core": [
+        "integration-libs/cdc/core/public_api"
+      ],
       "@spartacus/cdc": [
         "integration-libs/cdc/public_api"
       ],
